Refactoring and minor fixes
