+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/execution/bulkinsert/PreCombineRow.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi;
+package org.apache.hudi.execution.bulkinsert;
 
 import org.apache.spark.sql.Row;
